Add files via upload
SQL Project: Data Cleaning on Layoffs Dataset

This repository contains SQL scripts for cleaning and preprocessing a dataset on layoffs in 2022. The project includes steps for removing duplicates, standardizing data entries, handling null values, and performing necessary transformations to prepare the data for analysis. The original dataset is preserved for reference in a staging table, ensuring data integrity throughout the cleaning process.